// src/frontend/src/typings/spotify-api.d.ts
declare namespace SpotifyApi {
// ...
  interface CurrentPlaybackResponse
    extends CurrentlyPlayingObject,
      PlaybackObject {}
// ...
  interface ImageObject {
    height?: number;
    url: string;
    width?: number;
  }
// ...
  interface PlaybackObject {
    shuffle_state: boolean;
    repeat_state: PlaybackRepeatState;
  }

  interface CurrentlyPlayingObject {
    timestamp: number;
    device: UserDevice;
    progress_ms: number | null;
    is_playing: boolean;
    item: TrackObjectFull | null;
    context: ContextObject | null;
  }
// ...
}